+++ test/instagram.test.ts
@@ -4,7 +4,7 @@ import { getFollowerCount } from "../src"
 test("instagram", async () => {
   const count = await getFollowerCount({
     type: "instagram",
-    username: "rachel__0106",
+    username: "hello",
     auth: {
       username: "dummy_ig_user_egoist",
       password: import.meta.env.VITE_IG_PASSWORD,